[tests] Move typemap rewriter coverage to net11
The net10.0 host-test project cannot reference the net11.0 typemap
generator. Move the two generator-dependent remapping cases to the
existing net11.0 integration suite, preserving UTF-8 data and warning
assertions, and remove the legacy project's generator import.
